Update to recent distros and collate install steps
Provide a single script that is capable of performing the required
install steps needed to test various distros. This can be used by other
users in the future to simplify the steps required to have a successful
deployment.

Migrate the Vagrantfile used to be managed outside and uploaded as a
direct file, as that facilitates checking directly that the provided
config is valid.