Trigger tests on submodule update
Add the basic structure required to trigger testing all distributions
currently supported when an update is proposed to the .gitmodules file.

This can be used as a proxy to pick up latest updates from
vagrant-libvirt on a daily basis via dependabot in the hope of
maintaining a set of working instructions for each distribution.