Validate installation of vagrant-libvirt on multiple distros (#5)
Add the basic structure required to trigger testing all distributions
currently supported when an update is proposed to the .gitmodules file.

Use docker containers for the distros to validate linking of libraries
allow the vagrant bindings to work as expected to use libvirt from the
distro to launch the VM with the necessary limitation that it will not
be a perfect check. Using qemu to spin up the distros for a more
accurate check is not possible due to the performance overhead as
github actions do not allow for nested virtualization.

This can be used as a proxy to pick up latest updates from
vagrant-libvirt on a daily basis via dependabot in the hope of
maintaining a set of working instructions for each distribution.